+++ components/TheFooter.vue
@@ -3,35 +3,13 @@
     <div class="container">
       <div class="row">
         <div class="col-md-6 col-lg-6 col-12">
-          <h4 class="block-heading">Về chúng tôi</h4>
+          <h4 class="block-heading" >Về chúng tôi</h4>
           <ul>
-            <li class="mt-2">
-              <a
-                href="https://drive.google.com/file/d/1u4UuL7NfYHgN5X0c5g527VVz2-aiBNFX/view?usp=sharing"
-                class="link"
-                target="_blank"
-                >Quyết định công nhận Quỹ</a
-              >
-            </li>
-            <li class="mt-2">
-              <a
-                href="https://drive.google.com/file/d/10uBvIY73u0_b1e9vpv4h3Pi9BzOKEuni/view?usp=sharing"
-                class="link"
-                target="_blank"
-                >Quyết định thành lập Quỹ</a
-              >
-            </li>
-            <li class="mt-2">
-              <a
-                href="https://drive.google.com/file/d/1pFvjcXAoTR-NXC6nJ9ydE113T71D35bG/view?usp=sharing"
-                class="link"
-                target="_blank"
-                >Điều lệ Quỹ</a
-              >
+            <li class="mt-2" v-for="(menu, index) in menus" :key='index'>
+              <MenuPDF :id='menu.id' :name='menu.name' :link='menu.link' />
             </li>
           </ul>
         </div>
-
         <div class="col-12 col-lg-6">
           <p class="d-i-block mt-2">
             Copyright © 2021 Quỹ phát triển tài năng Việt. All rights reserved.
@@ -43,8 +21,48 @@
 </template>
 
 <script>
+
+import MenuPDF from "@/components/MenuPDF";
+
+// var loadingTask = pdf.createLoadingTask('/pdf/pdf_3.pdf');
+
 export default {
-  name: "AppFooter"
+  name: "AppFooter", 
+  components: {
+    MenuPDF
+  },
+  data() {
+    return {
+      // src: loadingTask,
+      // numPages: undefined,
+      menus: [
+        {
+          id: 'modal-1',
+          name: 'Quyết định công nhận Quỹ',
+          link: '/pdf/pdf_1.pdf',
+          page: 1
+        },
+        {
+          id: 'modal-2',
+          name: 'Quyết định thành lập Quỹ',
+          link: '/pdf/pdf_2.pdf',
+          page: 2
+        },
+        {
+          id: 'modal-3',
+          name: 'Điều lệ Quỹ',
+          link: '/pdf/pdf_3.pdf',
+          page: 12
+        }
+      ]
+    }
+  },
+  mounted() {
+
+  },
+  methods: {
+    
+  }
 };
 </script>
 
@@ -52,4 +70,7 @@ export default {
 .social-icon {
   width: 4rem;
 }
+// ul li {
+//   cursor: pointer;
+// }
 </style>