+++ pages/fund-info.vue
@@ -21,7 +21,7 @@
                           :class="{ active: tabActive === 1 }"
                           @click="setTabActive(1)"
                         >
-                          Cá nhân
+                          Cá nhân / Tổ chức
                         </div>
 
                         <div
@@ -173,8 +173,7 @@
         </div>
       </div>
     </div>
-    <div class='banner-bottom'>
-    </div>
+    <div class="banner-bottom"></div>
   </div>
 </template>
 
@@ -210,7 +209,7 @@ export default {
         data = data.filter((d) => this.normalizeee(d[this.tabActive === 1 ? 'fullname' : 'name']).includes(this.normalizeee(this.searchKey)))
       }
       if (this.sortBy == 'amount') {
-        data = data.sort((a, b) => a.amount >= b.amount ? this.sortValue === 'asc' ? -1 : 1 : this.sortValue === 'asc' ? 1 : -1)
+        data = data.sort((a, b) => parseInt(a.amount) >= parseInt(b.amount) ? this.sortValue === 'asc' ? -1 : 1 : this.sortValue === 'asc' ? 1 : -1)
       } else if (this.sortBy == 'updated_at') {
         data = data.sort((a, b) => new Date(a.updated_at).getTime() >= new Date(b.updated_at).getTime() ? this.sortValue === 'asc' ? -1 : 1 : this.sortValue === 'asc' ? 1 : -1)
       }
@@ -238,13 +237,13 @@ export default {
         if (this.tabActive == 1) {
           if (!this.customerLoaded) {
             this.isLoading = true;
-            await this.fetchRankPerson({perpage: 10000})
+            await this.fetchRankPerson({page: 0, perpage: 10000})
             this.customerLoaded = true
           }
         } else {
           if (!this.shopLoaded) {
             this.isLoading = true;
-            await this.fetchRankStore({perpage: 10000})
+            await this.fetchRankStore({page: 0, perpage: 10000})
             this.shopLoaded = true
           }
         }
@@ -325,16 +324,13 @@ export default {
   }
 }
 .tab {
-  width: 148px;
-  padding: 10px 34px;
+  padding: 10px;
   font-size: 17px;
   font-weight: bold;
   font-family: "Roboto Condensed", sans-serif;
   letter-spacing: 0;
   text-align: center;
-  //   &.tab-store {
-  //     margin: 0 24px;
-  //   }
+  flex: 1;
   &:hover {
     cursor: pointer;
   }
@@ -356,18 +352,15 @@ export default {
   font-family: "Roboto Condensed", sans-serif;
   font-size: 17px;
   .btn-sort {
-    width: 24px;
-    height: 24px;
-    border: 1px solid;
     outline: none;
     padding: 0;
     display: inline-block;
     text-align: center;
-    line-height: 19px;
+    line-height: 20px;
     cursor: pointer;
     img {
       max-width: 100%;
-      width: 16px;
+      width: 14px;
     }
     &:focus {
       outline: none;
@@ -383,6 +376,7 @@ export default {
   @media (min-width: 992px) {
     text-align: right;
     margin-top: 0;
+    padding-left: 0;
   }
 }